Start separating API from internals more strictly (Typed)
